+++ Servises/GeoCoder/GeoLocation.cs
@@ -4,8 +4,8 @@
     {
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Latitude { get; set; }
-        public double Longitude { get; set; }
+        public float Latitude { get; set; }
+        public float Longitude { get; set; }
 
     }
 }
